requirements.txt
discord==2.3.2
langchain==0.0.184
openai==1.9.0
openapi-schema-pydantic==1.2.4
pandas==2.2.0
PyPDF2==3.0.1
python-dotenv==1.0.0
requests @ file:///croot/requests_1690400202158/work
scikit-learn==1.4.0
scipy==1.12.0